bot/requirements.txt
langchain-google-genai==2.0.7
python-dotenv>=0.21.1
gunicorn==23.0.0
botbuilder-integration-aiohttp>=4.16.1
